Reorganized Splash into components folder
