+++ web/ugahacks-7/components/Splash/Countdown/index.ts
@@ -1,0 +1,3 @@
+import Countdown from "./Countdown";
+
+export default Countdown;
