+++ web/ugahacks-7/components/Splash/index.ts
@@ -1,0 +1,3 @@
+import Splash from "./Splash";
+
+export default Splash;
