+++ web/ugahacks-7/pages/index.tsx
@@ -1,59 +1,20 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import styles from '../styles/Home.module.css'
-import Splash from './splash'
+import { ReactElement } from "react";
+import Head from "next/head";
+import styles from "../styles/Home.module.css";
+import Splash from "../components/Splash";
 
-export default function Home() {
+export default function Home(): ReactElement {
   return (
-    <div> 
-    <Splash/>
     <div className={styles.container}>
       <Head>
         <title>UGAHacks 7</title>
-        <meta name="description" content="UGAHacks 7 Event Site [[ event info here ]]" />
+        <meta
+          name="description"
+          content="UGAHacks 7 Event Site [[ event info here ]]"
+        />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-            Welcome to 7.ugahacks.com!
-        </h1>
-
-        <p className={styles.description}>
-          <code className={styles.code}>This page is currently under construction!</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://ugahacks.com" className={styles.card}>
-              <h2>UGAHacks.com &rarr;</h2>
-              <p>UGAHacks&apos; organization site</p>
-          </a>
-
-          <a href="https://ugeorgia.ca1.qualtrics.com/jfe/form/SV_82Gn6acN70D0wg6" className={styles.card}>
-            <h2>UGAHacks 7&rarr;</h2>
-            <p>Pre-Register for our seventh event, coming this October 22-24th, 2021!</p>
-          </a>
-          
-          <a href="https://6.ugahacks.com" className={styles.card}>
-            <h2>6.ugahacks.com &rarr;</h2>
-            <p>Previous event site</p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
+      <Splash />
     </div>
-    </div> //remove me
-  )
+  );
 }
